+++ packages/cli-core/src/docker/index.js
@@ -1,6 +1,0 @@
-//
-// Copyright 2020 DXOS.org
-//
-
-export * from './container';
-export * from './image';