+++ packages/cli-core/src/shims.d.ts
@@ -1,0 +1,8 @@
+//
+// Copyright 2020 DXOS.org
+//
+
+declare module '@dxos/config';
+declare module 'console.table';
+declare module 'node-yaml';
+declare module 'yargs-unparser';